requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
imageio==2.34.0
importlib_resources==6.4.0
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
SimpleITK==2.3.1
six==1.16.0
tifffile==2023.7.10
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.1
